make `brainpy.tools.checking` as an independent module `brainpy.check`
