+++ brainpy/errors.py
@@ -1,6 +1,10 @@
 # -*- coding: utf-8 -*-
 
 
+
+
+
+
 class BrainPyError(Exception):
   """General BrainPy error."""
   pass
@@ -114,4 +118,59 @@ class ConcretizationTypeError(Exception):
       'as arguments, not keyword arguments, like "jit_f(v1, v2)" [<- wrong]. '
       'Please write it as "jit_f(static_k1=v1, static_k2=v2)" [<- right].'
     )
+
+
+_BRAINPYLIB_MINIMAL_VERSION = '0.1.3'
+
+
+try:
+  import jaxlib
+  del jaxlib
+except ModuleNotFoundError:
+  raise ModuleNotFoundError(
+    '''
+
+BrainPy needs jaxlib, please install it. 
+
+1. If you are using Windows system, install jaxlib through
+
+   >>> pip install jaxlib -f https://whls.blob.core.windows.net/unstable/index.html
+
+2. If you are using macOS platform, install jaxlib through
+
+   >>> pip install jaxlib -f https://storage.googleapis.com/jax-releases/jax_releases.html
+
+3. If you are using Linux platform, install jaxlib through
+
+   >>> pip install jaxlib -f https://storage.googleapis.com/jax-releases/jax_releases.html
+
+4. If you are using Linux + CUDA platform, install jaxlib through
+
+   >>> pip install jaxlib -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
+
+Note that the versions of "jax" and "jaxlib" should be consistent, like "jax=0.3.14" and "jaxlib=0.3.14".  
+
+For more detail installation instructions, please see https://brainpy.readthedocs.io/en/latest/quickstart/installation.html#dependency-2-jax 
+
+    ''') from None
+
+
+try:
+  import brainpylib
+
+  if brainpylib.__version__ < _BRAINPYLIB_MINIMAL_VERSION:
+    raise PackageMissingError(
+      f'\nbrainpy need "brainpylib>={_BRAINPYLIB_MINIMAL_VERSION}". \n'
+      f'Please install it through:\n\n'
+      f'>>> pip install brainpylib -U'
+    )
+
+  del brainpylib
+except ModuleNotFoundError:
+  raise PackageMissingError(
+    f'\nbrainpy need "brainpylib>={_BRAINPYLIB_MINIMAL_VERSION}". \n'
+    f'Please install "brainpylib>={_BRAINPYLIB_MINIMAL_VERSION}" through:\n\n'
+    f'>>> pip install brainpylib'
+  )
+
