+++ brainpy/tools/__init__.py
@@ -2,4 +2,7 @@
 
 from .codes import *
 from .others import *
-from .errors import *
+from .dicts import *
+from .others import *
+from .numba_util import *
+from .math_util import *
