--- brainpy/tools/others/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .dicts import *
-from .others import *
-from .numba_util import *
-from .math_util import *
